+++ Engine/src/Platform/Vulkan/Internal/VulkanCore.h
@@ -3,7 +3,6 @@
 
 #include <memory>
 #include <vector>
-#include <optional>
 
 #include <SDL3/SDL.h>
 #include <vulkan/vulkan.h>
@@ -11,26 +10,6 @@
 #include "Platform/Vulkan/Internal/VulkanDevice.h"
 
 namespace aero3d {
-
-struct QueueFamilyIndices 
-{
-    std::optional<uint32_t> GraphicsFamily;
-    std::optional<uint32_t> PresentFamily;
-
-    bool IsComplete() {
-        return GraphicsFamily.has_value() && PresentFamily.has_value();
-    }
-};
-
-struct VulkanPhysicalDevice
-{
-    VkPhysicalDevice Device;
-    VkPhysicalDeviceProperties Properties;
-    VkPhysicalDeviceFeatures Features;
-    VkPhysicalDeviceMemoryProperties MemoryProperties;
-    std::vector<VkQueueFamilyProperties> QueueFamilyProperties;
-    QueueFamilyIndices QueueFamilyIndices;
-};
 
 struct SwapChainSupportDetails 
 {
@@ -55,14 +34,9 @@ public:
     VkInstance GetInstance() { return m_Instance; }
     VkSurfaceKHR GetSurface() { return m_Surface; }
 
-    const VulkanPhysicalDevice& GetPhysicalDevice() { return m_PhysDevices[m_CurrentPhysDevice]; }
-    unsigned int GetCurrentPhysDeviceIndex() { return m_CurrentPhysDevice; }
-    void SetPhysicalDevice(unsigned int index) { m_CurrentPhysDevice = index; }
-
 private:
     bool CreateInstance();
     bool CreateSurface();
-    bool CreatePhysicalDevices();
     bool CreateSwapchain();
     bool CreateImageViews();
     bool CreateRenderPass();
@@ -79,9 +53,6 @@ private:
     VkInstance m_Instance;
     VkSurfaceKHR m_Surface;
     VkSwapchainKHR m_Swapchain;
-
-    unsigned int m_CurrentPhysDevice;
-    std::vector<VulkanPhysicalDevice> m_PhysDevices;
 
     VulkanDevice m_Device;
