+++ Engine/src/Platform/Vulkan/Internal/VulkanDevice.cpp
@@ -3,12 +3,14 @@
 #include <set>
 
 #include "Platform/Vulkan/Internal/VulkanUtils.h"
-#include "Platform/Vulkan/Internal/VulkanCore.h"
+#include "Utils/Assert.h"
+#include "Utils/Log.h"
 
 namespace aero3d {
 
 VulkanDevice::VulkanDevice()
-    : m_Device(VK_NULL_HANDLE), m_GraphicsQueue(VK_NULL_HANDLE), m_PresentQueue(VK_NULL_HANDLE)
+    : m_Device(VK_NULL_HANDLE), m_GraphicsQueue(VK_NULL_HANDLE), m_PresentQueue(VK_NULL_HANDLE),
+    m_Instance(VK_NULL_HANDLE), m_CurrentPhysDevice(0)
 {
 }
 
@@ -16,9 +18,78 @@ VulkanDevice::~VulkanDevice()
 {
 }
 
-bool VulkanDevice::Init()
+bool VulkanDevice::Init(VkInstance instance, VkSurfaceKHR surface)
 {
-    VulkanPhysicalDevice physDevice = g_VulkanCore->GetPhysicalDevice();
+    m_Instance = instance;
+
+    SearchForPhysDevices(surface);
+    CreateLogicalDevice();
+
+    return true;
+}
+
+void VulkanDevice::Shutdown()
+{
+    vkDestroyDevice(m_Device, nullptr);
+}
+
+void VulkanDevice::SearchForPhysDevices(VkSurfaceKHR surface)
+{
+    uint32_t physDeviceCount;
+    A3D_CHECK_VKRESULT(vkEnumeratePhysicalDevices(m_Instance, &physDeviceCount, nullptr));
+
+    if (!physDeviceCount)
+    {
+        LogErr(ERROR_INFO, "Failed to Enum PhysDevices.");
+    }
+
+    std::vector<VkPhysicalDevice> vkDevices(physDeviceCount);
+    m_PhysDevices.reserve(physDeviceCount);
+
+    A3D_CHECK_VKRESULT(vkEnumeratePhysicalDevices(m_Instance, &physDeviceCount, vkDevices.data()));
+
+    for (auto& vkDevice : vkDevices) {
+        VulkanPhysicalDevice physDev{};
+        physDev.Device = vkDevice;
+
+        vkGetPhysicalDeviceProperties(vkDevice, &physDev.Properties);
+        vkGetPhysicalDeviceFeatures(vkDevice, &physDev.Features);
+        vkGetPhysicalDeviceMemoryProperties(vkDevice, &physDev.MemoryProperties);
+
+        uint32_t queueCount = 0;
+        vkGetPhysicalDeviceQueueFamilyProperties(vkDevice, &queueCount, nullptr);
+
+        physDev.QueueFamilyProperties.resize(queueCount);
+        vkGetPhysicalDeviceQueueFamilyProperties(vkDevice, &queueCount, physDev.QueueFamilyProperties.data());
+
+        for (uint32_t i = 0; i < physDev.QueueFamilyProperties.size(); i++)
+        {
+            if (physDev.QueueFamilyProperties[i].queueFlags & VK_QUEUE_GRAPHICS_BIT)
+            {
+                physDev.QueueFamilyIndices.GraphicsFamily = i;
+            }
+
+            VkBool32 presentSupport = false;
+            vkGetPhysicalDeviceSurfaceSupportKHR(vkDevice, i, surface, &presentSupport);
+
+            if (presentSupport)
+            {
+                physDev.QueueFamilyIndices.PresentFamily = i;
+            }
+        }
+
+        if (physDev.QueueFamilyIndices.IsComplete())
+        {
+            m_PhysDevices.emplace_back(physDev);
+        }
+    }
+
+    Assert(ERROR_INFO, !m_PhysDevices.empty(), "Unnable to Find Suitable Physical Device.");
+}
+
+void VulkanDevice::CreateLogicalDevice()
+{
+    VulkanPhysicalDevice& physDevice = m_PhysDevices[m_CurrentPhysDevice];
 
     std::vector<VkDeviceQueueCreateInfo> queueCreateInfos;
     std::set<uint32_t> uniqueQueueFamilies = {
@@ -57,18 +128,10 @@ bool VulkanDevice::Init()
     if (m_Device == nullptr)
     {
         LogErr(ERROR_INFO, "Failed to create logical device.");
-        return false;
     }
 
     vkGetDeviceQueue(m_Device, physDevice.QueueFamilyIndices.GraphicsFamily.value(), 0, &m_GraphicsQueue);
     vkGetDeviceQueue(m_Device, physDevice.QueueFamilyIndices.PresentFamily.value(), 0, &m_PresentQueue);
-
-    return true;
-}
-
-void VulkanDevice::Shutdown()
-{
-    vkDestroyDevice(m_Device, nullptr);
 }
 
 } // namespace aero3d
